allow user to set empty dcs install dir
This will allow expert users to disable the automatic MissionScripting.lua replacement. There are many warnings and errors which have to be ignored to achieve this because DCS Liberation will not work with unmodified MissionScripting.lua

+++ qt_ui/liberation_install.py
@@ -112,7 +112,7 @@ def replace_mission_scripting_file():
     )
     liberation_scripting_path = "./resources/scripts/MissionScripting.lua"
     backup_scripting_path = "./resources/scripts/MissionScripting.original.lua"
-    if os.path.isfile(mission_scripting_path):
+    if install_dir != "" and os.path.isfile(mission_scripting_path):
         with open(mission_scripting_path, "r") as ms:
             current_file_content = ms.read()
         with open(liberation_scripting_path, "r") as libe_ms:
@@ -133,5 +133,9 @@ def restore_original_mission_scripting():
     )
     backup_scripting_path = "./resources/scripts/MissionScripting.original.lua"
 
-    if os.path.isfile(backup_scripting_path) and os.path.isfile(mission_scripting_path):
+    if (
+        install_dir != ""
+        and os.path.isfile(backup_scripting_path)
+        and os.path.isfile(mission_scripting_path)
+    ):
         copyfile(backup_scripting_path, mission_scripting_path)
